secureboot.config: add additional config options for USE=secureboot
- kexec allows us to replace the running kernel, without CONFIG_KEXEC_SIG this
allows us to effectively bypass secureboot.

- CONFIG_SECURITY_LOCKDOWN_LSM is designed to prevent altering the running
kernel image. Integrity mode is enabled automatically if this option is set and
the kernel is booted with secure boot enabled. So we select this mode as our
default for users who explicitly opt-in with USE=secureboot

- CONFIG_INTEGRITY_PLATFORM_KEYRING makes the kernel store the platform keys in
a separate keyring, CONFIG_INTEGRITY_MACHINE_KEYRING does the same for the MOK
list. This allows using these keys for CONFIG_KEXEC_SIG or to use the MOK list
keys to sign additional out-of-tree modules with a different key then the
in-tree modules (useful in pre-signed gentoo-kernel-bin).

Signed-off-by: Andrew Ammerlaan <andrewammerlaan@gentoo.org>
Closes: https://github.com/projg2/gentoo-kernel-config/pull/8
Signed-off-by: Michał Górny <mgorny@gentoo.org>